Immproved some UIUX
